Refactor: Move test mocks into a separate folder
We don't want to push the test mocks to Apps Script.
The mocks are for mimicking the Google implementation of the libraries available to Apps Script.

This puts the mock file into a separate `__mocks__` folder.
It also adds this folder to the list of files that clasp SHOULD NOT push to the Apps Script environment.