Use Promise.all to resolve all potential S3 message in the SQS Event
Given the structure of the SQS message and SQS Events,
there could be multiple events in an SQS message and multiple records in an S3 event.

We don't anticipate multiple events as we only send a message to SQS when an object is created in S3.
But with these changes we take into consideration the potential of an SQS message containing multiple messages,
so that we don't ignore data that should be processed by the lambda.

This adds code that puts all the `potential` S3 events into an array,
then calls `Promise.all` to resolve processing each event i.e return the data as a JSON object,
Then end result should be an array of JSON objects each representing a form submission.

Co-authored-by: Lawrence <acodeninja@users.noreply.github.com>