Rename lambda file and main function
This changes the name of the lambda file to `main` and the main function to `handler`.

This also update the serverless config to find the lambda file.

When we tested this code in devscratch we got errors such as:

- `Runtime.ImportModuleError: Error: Cannot find module`
- `handler is undefined or not exported`

Troubleshooting in the devscratch led us to try renaming the function and file based on what AWS would generate if you created the lambda in the console.

But we've come to understand that it was probably our serverless setup and file system setup that was not allowing the lambda to be found during the devscratch test.

This makes changes based on what we found from testing in a different environment and how it might apply here.