Test document is generated
This adds a test that checks that the  #createDocumentFromTemplate method is called with the generated auth token, retrieved S3 data, template ID and a title.

This also updates the #getDataFromS3 method to return an array of type `Record<string, string[]>` as the object is the same format as what is returned when we
call namedValues on an event in Apps Script.

This also updates the #createDocumentFromTemplate method to accept this kind of input.