Add fixes from testing lambda in test environment
- Format private key
This formats the private key of the service account. The reason for this is because when adding the key to the environment variables it does not format the new line characters and the auth fails.

- Add `Promise.all`
This encapsulates the code that loops through each form data object within a Promise.all. This means that the code will wait for all the await functions to resolve before exiting the lambda. Without these, the lambda doesn't wait for the entire doc generation to finish and exits prematurely.

- Add timeout of 30 seconds to the lambda.
The default timeout is 6 seconds. This updates it so that the lambda is up for 30 seconds before terminating itself.

Co-authored-by: Miles-Alford <miles.alford@hackney.gov.uk>