Change executor for deploy lambda job
The job currently fails with this error: `@social-care-referral-form-ingestion/referral-form-data-process@1.0.0: The engine "node" is incompatible with this module. Expected version ">=14.18.0". Got "14.17.5"`

We've come across this before error and I think when we use `aws-cli/default` as the executor, the version of node it has does not match our node engine requirements.