Add `~true` to ssm parameter call
During the deploy stage, `serverless` could not find the secrets for the API endpoint and API key.

From reading, when trying to access a secret in AWS, it seems we need to add `~true` to the end of the call in the `serverless` config.

We have `~true` set up already for the other environment variables and `serverless` has not warned about these ones.

This adds `~true` to try and fix `serverless` not getting the values for the API endpoint and key.