Generate document title based on number of referrals in the form
The referral process generates a Google Document that a member of MASH would use.
This happens once the MASH form is submitted.

Depending on the number of other people in a submitted referral,
this will generate a title for the generated google doc.

This copied the code used in the `sendDataToAPI` file to map the form data into an object.
As part of this commit, we introduce two new folders:
- helpers
- interfaces

The copied code has been refactored into separate functions for creating a list of clients and mapping the raw form data into an object.

Since the original code was copied from the `sendDataToAPI` file, in a separate comiit, we will refactor this file to import these functions as well.