Loop through form data to create resident details
This takes the form data and loops through to get each child's data.

The data is then used to create an object and this object is added to a list.

Once all the child information in a form has been mapped to a object and added to the list.

The list is sent as a part of the API request to add the data to the database.